backend/requirements/base.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-jazzmin==3.0.1
Pillow==10.4.0
python-decouple==3.8
python-slugify==8.0.4
whitenoise==6.7.0
gunicorn==22.0.0
